Take sys/vxworks/{fd,fs,io} from sys/unix instead.
